MERGED

+++ src/SalesManager.java
@@ -15,8 +15,8 @@ public class SalesManager {
         return max;
     }
 
-    public int min() {
-        int min = sales[0];
+    public long min() {
+        long min = sales[0];
         for (int i = 0; i < sales.length; i++) {
             if (sales[i] < min) {
                 min = sales[i];
@@ -24,7 +24,7 @@ public class SalesManager {
         }
         return min;
     }
-    public int average(){
+    public long average(){
         int average=0;
         for (int i = 0; i < sales.length; i++) {
             average+=sales[i];
